+++ src/Files.Uwp/ServicesImplementation/LocalizationService.cs
@@ -1,4 +1,4 @@
-﻿using Files.Backend.Services;
+﻿using Files.Shared.Services;
 using Microsoft.Toolkit.Uwp;
 
 namespace Files.Uwp.ServicesImplementation
